refactor: Refacto script structure
